+++ core/class/alarme_IMA.class.php
@@ -843,7 +843,7 @@ class alarme_IMA extends eqLogic {
     $response='';
 	try {
       	$eqlogic = eqLogic::byId($input);
-      	$imaProtectAPI = new imaProtectNewAPI($eqlogic->getConfiguration('login_ima'),$eqlogic->getConfiguration('password_ima'),$eqlogic->getConfiguration('cfgContactList'),$input);
+      	$imaProtectAPI = new imaProtectNewAPI($eqlogic->getConfiguration('login_ima'),$eqlogic->getConfiguration('password_ima'),$eqlogic->getConfiguration('cfgContactList'),$input,$eqlogic->getConfiguration('checkPwdXO'));
 		
 		if (!($imaProtectAPI->getDatasSession())) {
 			log::add('alarme_IMA', 'debug',  "	* Validation couple user / mdp");
@@ -868,6 +868,12 @@ class alarme_IMA extends eqLogic {
 	try {
 		$myImaProtectAlarm = $this->getInstanceIMAApi();
 	    log::add('alarme_IMA', 'debug',  "	* Extinction alarme");
+
+		$checkPwdXO=$this->getConfiguration('checkPwdXO');
+		if ($checkPwdXO == '1' && empty($pwd)) {
+			$this->manageErrorAPI('setAlarmToOff','Le code XO est nécessaire pour désarmer l\'alarme');		
+		}
+	
 		$myImaProtectAlarm->setAlarmToOff($pwd);
 	} catch (Exception $e) {
 	  $this->manageErrorAPI("setAlarmToOff",$e->getMessage());
@@ -960,9 +966,8 @@ class alarme_IMA extends eqLogic {
   
   private function getInstanceIMAApi(){
     try {
-      	$imaProtectAPI = new imaProtectNewAPI($this->getConfiguration('login_ima'),$this->getConfiguration('password_ima'),$this->getConfiguration('cfgContactList'),$this->getId());
+      	$imaProtectAPI = new imaProtectNewAPI($this->getConfiguration('login_ima'),$this->getConfiguration('password_ima'),$this->getConfiguration('cfgContactList'),$this->getId(),$this->getConfiguration('checkPwdXO'));
 				
-      	//if (!($imaProtectAPI->getContextFromTmpFile())) {
 		if (!($imaProtectAPI->getDatasSession())) {
 			log::add('alarme_IMA', 'debug',  "	* Validation couple user / mdp");
 			$imaProtectAPI->login();
@@ -1137,7 +1142,13 @@ class alarme_IMA extends eqLogic {
 				$replace['#' . $cmd->getLogicalId() . '_listValue#'] = $listOption;
 			}
 		}
-		  
+		 
+		//pass ima option for xo code
+		$replace['#checkPwdXO#'] = $this->getConfiguration('checkPwdXO');
+
+		//pass ima option if xo code is alphanumeric
+		$replace['#cfgXOAlpha#'] = $this->getConfiguration('cfgXOAlpha');
+
       log::add('alarme_IMA', 'debug',  "Function toHtml - Value replace : ".json_encode($replace));	
       $html = template_replace($replace, getTemplate('core', $_version, 'default_alarme_IMA', 'alarme_IMA'));
       cache::set('widgetHtml' . $_version . $this->getId(), $html, 1);